third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/connect-src-anchor-ping-expected.txt

CONSOLE ERROR: Connecting to 'https://localhost:8443/resources/dummy.txt' violates the following Content Security Policy directive: "connect-src http://127.0.0.1:8000". The action has been blocked.
